Consolidate kv-run-commands into kv-selectors
