+++ ui/tests/acceptance/secrets/backend/kv/kv-v2-workflow-create-test.js
@@ -11,8 +11,12 @@ import { setupApplicationTest } from 'vault/tests/helpers';
 import authPage from 'vault/tests/pages/auth';
 import { deleteEngineCmd, mountEngineCmd, runCmd, tokenWithPolicyCmd } from 'vault/tests/helpers/commands';
 import { personas } from 'vault/tests/helpers/kv/policy-generator';
-import { clearRecords, writeVersionedSecret } from 'vault/tests/helpers/kv/kv-run-commands';
-import { KV_FORM, KV_WORKFLOW } from 'vault/tests/helpers/kv/kv-selectors';
+import {
+  KV_FORM,
+  KV_WORKFLOW,
+  clearRecords,
+  writeVersionedSecret,
+} from 'vault/tests/helpers/kv/kv-selectors';
 import { grantAccessForWrite, setupControlGroup } from 'vault/tests/helpers/control-groups';
 import { GENERAL } from 'vault/tests/helpers/general-selectors';
 import { KV_METADATA_DETAILS } from 'vault/tests/helpers/components/kv/page/secret/metadata/details-selectors';
